Added a timeout on the sentence processor
And then I fixed some inefficiences I discovered while debugging what looked like an infinite
loop.  It was actually just a really deep and slow recursion, caused by calling `mkString` inside
of a string interpolation.  I'm really not sure what that was so slow, but removing that solved
the problem.


Former-commit-id: acbe259d00c66439ab4f773a1e7847e8ef58ee4d
Former-commit-id: 7b347d35dc70cf0ea8c9a19a1606b2ddb64ee24f