Major refactoring, SimpleLSTMSolver now appears to work
The code is now structured in proper modules, with a run script that is outside the module
hierarchy.  In addition, the NNSolver went through some major restructuring, making it cleaner and
easier to extend to new kinds of models, maximizing reusability of various components.

A major goal of this refactoring, other than just focusing on code health, was to make it easier
to change how the input is encoded, to use pre-trained word embeddings, or initializing an
embedding layer with word2vec.  That should be pretty easy now, by adding a new kind of Dataset
(i.e, EmbeddedDataset), constructed from an IndexedDataset.