Fixed a couple of bugs in the memory network solver
Background wasn't getting read right, because I forgot to convert the index to an int, and I
needed to convert a tuple to a list in MemoryNetworkSolver.prep_labeled_data().

But, it now works!  The refactoring is finished.  Next: making an EmbeddedDataset, where indices
get mapped to vectors, with a projection layer instead of an embedding layer.