Refactoring the Instance objects
The point of this refactoring is to make it easier to add a new type.  I want to add an Instance
type that has a question and associated answer options.  To do this, and still allow background
data to be associated with the instance, I made the BackgroundInstance use composition instead of
inheritance, and I made the code that reads data take an instance type.