Allowing memory-network-specific layers to be reused
This means saving them as member variables, both in the solver and in the entailment model.  This
is necessary in order to allow for pre-training, which requires the same layer objects to be used
for two different models.

I also took the opportunity to clean up the way arguments are passed to the entailment model
constructors.