Two main improvements: VectorMatrixMerge and separate embeddings (#114)
* Two main improvements: VectorMatrixMerge and separate embeddings

There were a couple of places where we had lots of code in _build_model() that does a fancy merge.
 I made that into its own layer, to clean up the code a bit, and allow for more code re-use,
because we did this kind of merge in a couple of different spots.

Also, in looking at how end-to-end memory networks are structured, I realized that we needed to
handle the embedding a little differently.  Now you can create as many different embeddings as you
like, by passing different embedding names to the _embed_input() function.  I don't know why you
would want to do this except in the bAbI dataset, but you now can if you want to...

* Fixed a couple of bugs found by pylint (thanks pylint!)

* Removed unnecessary print statements

* Made debug output handle the new embedding stuff

* Minor fixes from Pradeep's comments