Got dynamic padding actually working for sequence tagger (#319)
* Got dynamic padding actually working for sequence tagger

The main issue was with the CNN encoder used to get word encodings.  I
needed to make it agnostic to the input length, which it now is.

This led to a ~4x speedup in runtime, going from over 4000s per epoch
to ~1000s.  And that's an automatic gain, with a couple of flags, for
any model that's compatible.

* Added a uses_data_generators field to Trainer

* Moved uses_data_generators to be a method, not an attribute